refactor: Alteração tipo de dado Meta de Long para Float

+++ src/main/java/br/edu/fesa/vaievem/models/ContaBancaria.java
@@ -12,7 +12,7 @@ public class ContaBancaria implements Serializable {
     private String descricao;
     private String numeroAgencia;
     private String numeroConta;
-    private Long meta;
+    private float meta;
     
     // Atributos de relacionamento
     private Usuario usuario;
@@ -33,12 +33,12 @@ public class ContaBancaria implements Serializable {
         this.idContaBancaria = idContaBancaria;
         this.descricao = descricao;
     }
-    public ContaBancaria(Long idContaBancaria, String descricao, Long meta) {
+    public ContaBancaria(Long idContaBancaria, String descricao, float meta) {
         this.idContaBancaria = idContaBancaria;
         this.descricao = descricao;
         this.meta = meta;
     }
-    public ContaBancaria(Long idContaBancaria, String descricao, String numeroAgencia, String numeroConta, Long meta) {
+    public ContaBancaria(Long idContaBancaria, String descricao, String numeroAgencia, String numeroConta, float meta) {
         this.idContaBancaria = idContaBancaria;
         this.descricao = descricao;
         this.numeroAgencia = numeroAgencia;
@@ -75,10 +75,10 @@ public class ContaBancaria implements Serializable {
         this.numeroConta = numeroConta;
     }
     
-    public Long getMeta() {
+    public float getMeta() {
         return meta;
     }
-    public void setMeta(Long meta) {
+    public void setMeta(float meta) {
         this.meta = meta;
     }
